refactor: Capsules 도메인 분리 및 잔여 레거시 구조 제거 (Phase 3)
- Capsules 도메인 (Controller, Service, Repository) 도입
- src/app, src/schemas 폴더 내부 파일 완벽 이관 후 찌꺼기 파일 제거
