+++ src/modules/capsules/dto/capsules.command.dto.ts
@@ -1,0 +1,10 @@
+export type CapsuleMutationDto = {
+  slug?: string;
+  title?: string;
+  openAt?: string;
+};
+
+export type MessageInputDto = {
+  nickname?: string;
+  content?: string;
+};
