+++ src/modules/capsules/dto/delete-capsule.dto.ts
@@ -1,0 +1,9 @@
+import { z } from "../../../openapi/zod-extend";
+import { slugSchema } from "./shared.dto";
+
+export const deleteCapsuleResponseSchema = z
+  .object({
+    deleted: z.literal(true).openapi({ example: true }),
+    slug: slugSchema,
+  })
+  .openapi("DeleteCapsuleResponse");
