+++ src/modules/capsules/dto/index.ts
@@ -1,0 +1,9 @@
+export * from "./shared.dto";
+export * from "./capsules.command.dto";
+export * from "./slug-reservation.dto";
+export * from "./create-capsule.dto";
+export * from "./get-capsule.dto";
+export * from "./verify-capsule-password.dto";
+export * from "./update-capsule.dto";
+export * from "./delete-capsule.dto";
+export * from "./create-message.dto";
